\name{akima}
\alias{akima}
\docType{data}
\title{
Wave distortions data.}
\description{
Data taken from a study of waveform distortion in electronic circuits.
}
\usage{data("akima")}
\format{
  The format is:
 num [1:11, 1:9] 58.2 37.2 22.4 21.8 16.8 12 7.4 3.2 0 0 ...
}
\source{
Akima, H; 'A Method of Bivariate Interpolation and Smooth Surface Fitting Based on Local
Procedures', Comm. ACM, 17, 1974, pp 26-27
}

\examples{
data(akima)
}
\keyword{datasets}
\keyword{Akima}
