\name{spsi}
\docType{package}
\title{Shape Preserving Spline Interpolation}
\description{
Shape preserving Uni-variate and Bi-variate spline interpolation using method of polynomials of variable degree.
}
\details{
\tabular{ll}{
Package: \tab spsi\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2015-08-19\cr
License: \tab GPL-2 or greater\cr
LazyLoad: \tab yes\cr
}
}
\author{
Szymon Sacher <s1340144@sms.ed.ac.uk> & Andrew Clausen <andrew.clausen@ed.ac.uk> \cr
Excerpts adapted from Fortran code Copyright (C) Paolo Costantini


}
\keyword{spline}
\keyword{interpolation}
\seealso{
  \code{\link{sps_fun}}
  \code{\link{sps_prep}}
}
